.......... [DEV-2184] updated autotests in 6.0 branch

+++ src/libs/zbxserver/evalfunc.c
@@ -2824,7 +2824,7 @@ static int	evaluate_TREND(zbx_variant_t *value, DC_ITEM *item, const char *func,
 		season_processed = (int)((double)season / 3600);
 
 		ret = trends_eval_stl(table, item->itemid, start, end, start_detect_period, end_detect_period,
-				(int)season_processed, deviations, dev_alg, (int)s_window, &value_dbl, error);
+				season_processed, deviations, dev_alg, (int)s_window, &value_dbl, error);
 
 		zbx_free(dev_alg);
 	}
